refactor: split cluster sync helpers
